[05] 소스코드/bigdata/requirements.txt
asgiref==3.2.7
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
django-cors-headers==3.3.0
djangorestframework==3.11.0
drf-yasg==1.17.1
entrypoints==0.3
idna==2.9
importlib-metadata==1.5.0
inflection==0.4.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.0
Jinja2==2.11.2
# joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1589812474002/work
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
mkl-service==2.3.0
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.4
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-surprise==1.1.0
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
# tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1588513834538/work
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
wincertstore==0.2
zipp==3.1.0
